add support for CUDAtk12.9 (#7394)
CUDA Toolkit 12.9 has been out for a while. The build currently fails
when it is installed as the builder checks against hardcoded values.
this PR adds the value 12.9. a better mechanism would be to check
dynamically that the major number is the same... maybe next time when
CUDA13 comes out :)

Signed-off-by: LosCrossos <165311345+loscrossos@users.noreply.github.com>